+++ lib/liquid_tabbar_minimize.dart
@@ -1,3 +1,4 @@
 export 'src/liquid_bottom_navigation_bar.dart';
 export 'src/liquid_tab_item.dart';
 export 'src/liquid_tabbar.dart';
+export 'src/liquid_route_observer.dart';